requirements.txt
# TXD - by benzoXdev
Pillow
discord.py
pyautogui
pyinstaller
requests
pywin32
opencv-python
psutil
pyperclip